pyproject.toml
[tool.poetry]
name = "explainnn"
version = "1.0"
description = ""
authors = ["Anonymous"]

[tool.poetry.dependencies]
python = "3.9.7"
torch = "1.11.0"
torchvision = "0.12.0"
numpy = "1.19.5"
matplotlib = "3.3.4"
captum = "0.4.1"
scipy = "1.5.4"
PyYAML = "6.0"
ipykernel = "6.13.0"
jupyter = "1.0.0"
jupyter_nbextensions_configurator = "0.4.1"
jupyter-contrib-nbextensions = "0.5.1"
pytest = "6.2.5"
setuptools = "62.3.2"
wheel = "0.37.1"
torchray = "^1.0.0"
tensorflow = "2.6.2"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
